# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portfolio Manager API",
    "description": "Phase 1 backend. Not financial advice.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/portfolios/{portfolio_id}/imports": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Create Import",
        "operationId": "create_import_api_portfolios__portfolio_id__imports_post",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "description": "The raw image bytes (not multipart).",
          "content": {
            "image/png": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "image/jpeg": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            },
            "image/webp": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        }
      }
    },
    "/api/portfolios/{portfolio_id}/imports/rows": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Create Import From Rows",
        "description": "On-device OCR path: no image ever reaches the server. No OCR consent is needed (no third\nparty and no server-side reading).",
        "operationId": "create_import_from_rows_api_portfolios__portfolio_id__imports_rows_post",
        "parameters": [
          {
            "name": "portfolio_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Portfolio Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRowsBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/{draft_id}": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Get Import",
        "operationId": "get_import_api_imports__draft_id__get",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "imports"
        ],
        "summary": "Patch Import",
        "operationId": "patch_import_api_imports__draft_id__patch",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/imports/{draft_id}/confirm": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Confirm Import",
        "operationId": "confirm_import_api_imports__draft_id__confirm_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Draft Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportDraftOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "portfolio_id": {
            "type": "integer",
            "title": "Portfolio Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "confirmed",
              "discarded"
            ],
            "title": "Status"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ImportRowModel"
            },
            "type": "array",
            "title": "Rows"
          },
          "proposed_changes": {
            "items": {
              "$ref": "#/components/schemas/ProposedChange"
            },
            "type": "array",
            "title": "Proposed Changes"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "portfolio_id",
          "status",
          "rows",
          "proposed_changes",
          "expires_at"
        ],
        "title": "ImportDraftOut"
      },
      "ImportPatch": {
        "properties": {
          "rows": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ImportRowModel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rows"
          },
          "proposed_changes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProposedChange"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Changes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ImportPatch"
      },
      "ImportRowModel": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index",
            "default": 0
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "tase_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tase Number"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "ILS"
          },
          "unit": {
            "type": "string",
            "enum": [
              "ILS",
              "agorot",
              "USD"
            ],
            "title": "Unit",
            "default": "ILS"
          },
          "matched_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Name"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/MatchCandidate"
            },
            "type": "array",
            "title": "Candidates"
          },
          "flags": {
            "items": {
              "type": "string",
              "enum": [
                "missing_fields",
                "value_mismatch",
                "unmatched",
                "low_confidence_match",
                "currency_changed",
                "unit_mismatch"
              ]
            },
            "type": "array",
            "title": "Flags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ImportRowModel"
      },
      "ImportRowsBody": {
        "properties": {
          "rows": {
            "items": {
              "$ref": "#/components/schemas/ImportRowModel"
            },
            "type": "array",
            "maxItems": 200,
            "title": "Rows"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "rows"
        ],
        "title": "ImportRowsBody",
        "description": "On-device OCR: the browser parsed the screenshot, only the stock rows are sent."
      },
...
      "MatchCandidate": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name",
          "score"
        ],
        "title": "MatchCandidate",
        "description": "A possible security for a row whose name match was too weak to accept automatically."
      },
...
      "ProposedChange": {
        "properties": {
          "row_index": {
            "type": "integer",
            "title": "Row Index"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "type": {
            "type": "string",
            "enum": [
              "buy",
              "sell",
              "deposit",
              "withdrawal"
            ],
            "title": "Type"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "row_index",
          "symbol",
          "type",
          "quantity",
          "amount",
          "currency"
        ],
        "title": "ProposedChange"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}